# backend/docs/openapi.snapshot.json
{
  "components": {
    "schemas": {
...
      "DiscogsConnectIn": {
        "example": {
          "access_token": "discogs_access_token_redacted",
          "external_user_id": "discogs_user_2048",
          "token_metadata": {
            "scope": "identity wantlist inventory",
            "token_type": "oauth"
          }
        },
        "properties": {
          "access_token": {
            "anyOf": [
              {
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Access Token"
          },
          "external_user_id": {
            "maxLength": 120,
            "minLength": 1,
            "title": "External User Id",
            "type": "string"
          },
          "token_metadata": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Token Metadata"
          }
        },
        "required": [
          "external_user_id"
        ],
        "title": "DiscogsConnectIn",
        "type": "object"
      },
      "DiscogsConnectOut": {
        "example": {
          "connected": true,
          "connected_at": "2026-01-10T15:34:12.123456+00:00",
          "external_user_id": "discogs_user_2048",
          "provider": "discogs"
        },
        "properties": {
          "connected": {
            "title": "Connected",
            "type": "boolean"
          },
          "connected_at": {
            "format": "date-time",
            "title": "Connected At",
            "type": "string"
          },
          "external_user_id": {
            "title": "External User Id",
            "type": "string"
          },
          "provider": {
            "title": "Provider",
            "type": "string"
          }
        },
        "required": [
          "provider",
          "external_user_id",
          "connected",
          "connected_at"
        ],
        "title": "DiscogsConnectOut",
        "type": "object"
      },
      "DiscogsDisconnectIn": {
        "properties": {
          "revoke": {
            "default": true,
            "title": "Revoke",
            "type": "boolean"
          }
        },
        "title": "DiscogsDisconnectIn",
        "type": "object"
      },
      "DiscogsDisconnectOut": {
        "properties": {
          "disconnected": {
            "title": "Disconnected",
            "type": "boolean"
          },
          "provider": {
            "title": "Provider",
            "type": "string"
          }
        },
        "required": [
          "provider",
          "disconnected"
        ],
        "title": "DiscogsDisconnectOut",
        "type": "object"
      },
      "DiscogsImportIn": {
        "example": {
          "source": "both"
        },
        "properties": {
          "source": {
            "default": "both",
            "enum": [
              "wantlist",
              "collection",
              "both"
            ],
            "title": "Source",
            "type": "string"
          }
        },
        "title": "DiscogsImportIn",
        "type": "object"
      },
      "DiscogsImportJobOut": {
        "example": {
          "completed_at": "2026-01-20T10:00:07+00:00",
          "created_at": "2026-01-20T10:00:00+00:00",
          "created_count": 140,
          "error_count": 0,
          "errors": [],
          "id": "b9f9402e-6f7a-4ced-9ca8-a0f6306ee4ef",
          "import_scope": "both",
          "imported_count": 200,
          "page": 3,
          "processed_count": 200,
          "provider": "discogs",
          "started_at": "2026-01-20T10:00:00+00:00",
          "status": "completed",
          "updated_at": "2026-01-20T10:00:07+00:00",
          "updated_count": 60,
          "user_id": "8f2a5009-c0a2-4f90-8f1b-c1716c26bf06"
        },
        "properties": {
          "completed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "created_count": {
            "title": "Created Count",
            "type": "integer"
          },
          "cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cursor"
          },
          "error_count": {
            "title": "Error Count",
            "type": "integer"
          },
          "errors": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Errors"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "import_scope": {
            "title": "Import Scope",
            "type": "string"
          },
          "imported_count": {
            "title": "Imported Count",
            "type": "integer"
          },
          "page": {
            "title": "Page",
            "type": "integer"
          },
          "processed_count": {
            "title": "Processed Count",
            "type": "integer"
          },
          "provider": {
            "title": "Provider",
            "type": "string"
          },
          "started_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "updated_count": {
            "title": "Updated Count",
            "type": "integer"
          },
          "user_id": {
            "format": "uuid",
            "title": "User Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "user_id",
          "provider",
          "import_scope",
          "status",
          "cursor",
          "page",
          "processed_count",
          "imported_count",
          "created_count",
          "updated_count",
          "error_count",
          "errors",
          "started_at",
          "completed_at",
          "created_at",
          "updated_at"
        ],
        "title": "DiscogsImportJobOut",
        "type": "object"
      },
      "DiscogsImportedItemListOut": {
        "properties": {
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/DiscogsImportedItemOut"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "source": {
            "enum": [
              "wantlist",
              "collection"
            ],
            "title": "Source",
            "type": "string"
          }
        },
        "required": [
          "source",
          "limit",
          "offset",
          "count",
          "items"
        ],
        "title": "DiscogsImportedItemListOut",
        "type": "object"
      },
      "DiscogsImportedItemOut": {
        "properties": {
          "artist": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Artist"
          },
          "discogs_master_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Discogs Master Id"
          },
          "discogs_release_id": {
            "title": "Discogs Release Id",
            "type": "integer"
          },
          "open_in_discogs_url": {
            "title": "Open In Discogs Url",
            "type": "string"
          },
          "source": {
            "enum": [
              "wantlist",
              "collection"
            ],
            "title": "Source",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "watch_release_id": {
            "format": "uuid",
            "title": "Watch Release Id",
            "type": "string"
          },
          "year": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Year"
          }
        },
        "required": [
          "watch_release_id",
          "discogs_release_id",
          "discogs_master_id",
          "title",
          "artist",
          "year",
          "source",
          "open_in_discogs_url"
        ],
        "title": "DiscogsImportedItemOut",
        "type": "object"
      },
      "DiscogsOAuthCallbackIn": {
        "properties": {
          "code": {
            "minLength": 1,
            "title": "Code",
            "type": "string"
          },
          "state": {
            "minLength": 8,
            "title": "State",
            "type": "string"
          }
        },
        "required": [
          "state",
          "code"
        ],
        "title": "DiscogsOAuthCallbackIn",
        "type": "object"
      },
      "DiscogsOAuthStartIn": {
        "properties": {
          "scopes": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "minItems": 1,
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scopes"
          }
        },
        "title": "DiscogsOAuthStartIn",
        "type": "object"
      },
      "DiscogsOAuthStartOut": {
        "properties": {
          "authorize_url": {
            "title": "Authorize Url",
            "type": "string"
          },
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          },
          "provider": {
            "title": "Provider",
            "type": "string"
          },
          "scopes": {
            "items": {
              "type": "string"
            },
            "title": "Scopes",
            "type": "array"
          },
          "state": {
            "title": "State",
            "type": "string"
          }
        },
        "required": [
          "provider",
          "authorize_url",
          "state",
          "scopes",
          "expires_at"
        ],
        "title": "DiscogsOAuthStartOut",
        "type": "object"
      },
      "DiscogsOpenInDiscogsOut": {
        "properties": {
          "open_in_discogs_url": {
            "title": "Open In Discogs Url",
            "type": "string"
          },
          "source": {
            "enum": [
              "wantlist",
              "collection"
            ],
            "title": "Source",
            "type": "string"
          },
          "watch_release_id": {
            "format": "uuid",
            "title": "Watch Release Id",
            "type": "string"
          }
        },
        "required": [
          "watch_release_id",
          "source",
          "open_in_discogs_url"
        ],
        "title": "DiscogsOpenInDiscogsOut",
        "type": "object"
      },
      "DiscogsStatusOut": {
        "example": {
          "connected": true,
          "connected_at": "2026-01-10T15:34:12.123456+00:00",
          "external_user_id": "discogs_user_2048",
          "has_access_token": true,
          "provider": "discogs"
        },
        "properties": {
          "connected": {
            "title": "Connected",
            "type": "boolean"
          },
          "connected_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Connected At"
          },
          "external_user_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "External User Id"
          },
          "has_access_token": {
            "default": false,
            "title": "Has Access Token",
            "type": "boolean"
          },
          "provider": {
            "title": "Provider",
            "type": "string"
          }
        },
        "required": [
          "connected",
          "provider"
        ],
        "title": "DiscogsStatusOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "watch-wax-api",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/integrations/discogs/connect": {
      "post": {
        "operationId": "connect_discogs_api_integrations_discogs_connect_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DiscogsConnectIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiscogsConnectOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Connect Discogs",
        "tags": [
          "integrations",
          "discogs"
        ]
      }
    },
    "/api/integrations/discogs/disconnect": {
      "post": {
        "operationId": "disconnect_discogs_api_integrations_discogs_disconnect_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DiscogsDisconnectIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiscogsDisconnectOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Disconnect Discogs",
        "tags": [
          "integrations",
          "discogs"
        ]
      }
    },
    "/api/integrations/discogs/import": {
      "post": {
        "operationId": "import_discogs_api_integrations_discogs_import_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DiscogsImportIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiscogsImportJobOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Import Discogs",
        "tags": [
          "integrations",
          "discogs"
        ]
      }
    },
    "/api/integrations/discogs/import/{job_id}": {
      "get": {
        "operationId": "get_import_job_api_integrations_discogs_import__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiscogsImportJobOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Import Job",
        "tags": [
          "integrations",
          "discogs"
        ]
      }
    },
    "/api/integrations/discogs/imported-items": {
      "get": {
        "operationId": "list_imported_discogs_items_api_integrations_discogs_imported_items_get",
        "parameters": [
          {
            "in": "query",
            "name": "source",
            "required": true,
            "schema": {
              "enum": [
                "wantlist",
                "collection"
              ],
              "title": "Source",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 25,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiscogsImportedItemListOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Imported Discogs Items",
        "tags": [
          "integrations",
          "discogs"
        ]
      }
    },
    "/api/integrations/discogs/imported-items/{watch_release_id}/open-in-discogs": {
      "get": {
        "operationId": "open_imported_item_in_discogs_api_integrations_discogs_imported_items__watch_release_id__open_in_discogs_get",
        "parameters": [
          {
            "in": "path",
            "name": "watch_release_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Watch Release Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "source",
            "required": true,
            "schema": {
              "enum": [
                "wantlist",
                "collection"
              ],
              "title": "Source",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiscogsOpenInDiscogsOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Open Imported Item In Discogs",
        "tags": [
          "integrations",
          "discogs"
        ]
      }
    },
    "/api/integrations/discogs/oauth/callback": {
      "post": {
        "operationId": "complete_discogs_oauth_api_integrations_discogs_oauth_callback_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DiscogsOAuthCallbackIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiscogsConnectOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Complete Discogs Oauth",
        "tags": [
          "integrations",
          "discogs"
        ]
      }
    },
    "/api/integrations/discogs/oauth/start": {
      "post": {
        "operationId": "start_discogs_oauth_api_integrations_discogs_oauth_start_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DiscogsOAuthStartIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiscogsOAuthStartOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Start Discogs Oauth",
        "tags": [
          "integrations",
          "discogs"
        ]
      }
    },
    "/api/integrations/discogs/status": {
      "get": {
        "operationId": "discogs_status_api_integrations_discogs_status_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiscogsStatusOut"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Discogs Status",
        "tags": [
          "integrations",
          "discogs"
        ]
      }
    },
...
  }
}